django/requirements.txt
iptools
django
djangorestframework
django-rest-auth
mysql-python
django-debug-toolbar
drf-nested-routers=0.11.1
django-rest-assured
factory_boy